requirements.txt
Django==5.1.1
Pillow==10.4.0
django-debug-toolbar~=4.4.6
django-mptt~=0.16.0